definitions: 42.2/Tumbleweed: Increase disk size to 40G
The vagrant images can be used as base VM systems to run automated
tests on them. Sometimes, 20G is not enough to accomondate all the
downloaded packages so we double the disk size. 20G is also pretty
low since the root filesystem is on btrfs which uses a lot of space
for housekeeping work.