scripts: base: base.sh: Only install required packages
A 'zypper dup' on vagrant image leads to tons of packages being merged
because of the default 'solver.onlyRequires' option. The Vagrant images
are meant to be small and provide only what's absolutely necessary to
users so they can build on top of it. As such, lets just disable the
recommended packages by default.